Extract serialization logic
